sample/requirements.txt
psycopg2-binary==2.8.6
requests==2.24.0
